+++ infrastructure/restapi/src/main/java/tech/pegasys/teku/infrastructure/restapi/AuthorizationManager.java
@@ -16,22 +16,18 @@ package tech.pegasys.teku.infrastructure.restapi;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_UNAUTHORIZED;
 
 import io.javalin.http.Context;
-import io.javalin.http.Handler;
-import io.javalin.security.AccessManager;
-import io.javalin.security.RouteRole;
 import java.io.File;
 import java.io.IOException;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Set;
 import java.util.stream.Stream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import tech.pegasys.teku.infrastructure.exceptions.InvalidConfigurationException;
 
-public class AuthorizationManager implements AccessManager {
+public class AuthorizationManager {
   private static final Logger LOG = LogManager.getLogger();
   private static final String BEARER_PREFIX = "Bearer ";
   private final String bearer;
@@ -90,29 +86,27 @@ public class AuthorizationManager implements AccessManager {
     ctx.status(SC_UNAUTHORIZED);
   }
 
-  @Override
-  public void manage(final Handler handler, final Context ctx, final Set<? extends RouteRole> set)
-      throws Exception {
-    if (ctx.matchedPath().equals("/swagger-docs")) {
-      handler.handle(ctx);
-      return;
+  public boolean manage(final Context ctx) {
+    if (ctx.path().startsWith("/swagger-") || ctx.path().startsWith("/webjars")) {
+      return true;
     }
 
     if (bearer.isEmpty()) {
       unauthorized(ctx, "API Reject - no bearer loaded by server.");
-      return;
+      return false;
     }
 
     final String auth = ctx.header("Authorization");
     if (auth == null || !auth.startsWith(BEARER_PREFIX)) {
       unauthorized(ctx, "API Reject - authorization bearer missing from request header");
-      return;
+      return false;
     }
     if (!auth.substring(BEARER_PREFIX.length()).equals(bearer)) {
       unauthorized(ctx, "API Reject - Unauthorized");
-      return;
+      return false;
     }
+    return true;
 
-    handler.handle(ctx);
+    //    handler.handle(ctx);
   }
 }