+++ CHANGELOG.md
@@ -13,5 +13,6 @@ the [releases page](https://github.com/Consensys/teku/releases).
 
 ### Additions and Improvements
 - Improve block rewards calculation performance for `/eth/v3/validator/blocks/{slot}` block production beacon node API.
+- Updated Javalin to v.6 (used by rest-api and keymanager-api).
 
 ### Bug Fixes